updated book editing buttons, added user book viewer

+++ books/views.py
@@ -36,6 +36,17 @@ def list_books(request: HttpRequest):
         books = Book.objects.all().order_by("-title")
     return render(request, template_name="books/home.html", context={"books": books})
 
+def list_user_books(request: HttpRequest, user_pk: int):
+
+    sort = request.GET.get("sort")
+    books = Book.objects.filter(user_id=user_pk).order_by("pk")
+
+    if sort == "asc":
+        books = Book.objects.filter(user_id=user_pk).order_by("title")
+    if sort == "desc":
+        books = Book.objects.filter(user_id=user_pk).order_by("-title")
+    return render(request, template_name="books/home.html", context={"books": books})
+
 @login_required
 def create_book(request: HttpRequest):
     if request.method == "POST":
@@ -72,15 +83,18 @@ def update_book(request: HttpRequest, pk: int):
 
         return render(request, template_name="books/update_book_form.html", context={"form": form})
 
-
+@login_required
 def delete_book(request: HttpRequest, pk: int):
     book = get_object_or_404(Book, pk=pk)
 
-    if request.method == "POST":
-        book.delete()
-        return redirect("home")
+    if request.user.pk == book.user.pk:
+        if request.method == "POST":
+            book.delete()
+            return redirect("home")
+        else:
+            return render(request, template_name="books/book_confirm_delete.html", context={"book": book})
     else:
-        return render(request, template_name="books/book_confirm_delete.html", context={"book": book})
+        return HttpResponse("You are not authorized to delete another user's book")
